Show blank square if no image uploaded
This commit modifies the person partial to return an empty div if the
person has no image uploaded and adds the CSS to style the empty div
as a blank square that behaves like an image.

Something to note is the magic number percentage that is needed in order
to mantain the same aspect ratio as the images.