Set `X-Frame-Options` header to `ALLOWALL`
Rails 4 starts setting a default `X-Frame-Options` policy of
`SAMEORIGIN` which prevents this application from being
displayed in an iframe outside of GOV.UK, for example in
the side-by-side browser.

As this application only contains information pages and nothing
transactional, we have decided to allow it to be included in
iframes on external sites.

This is explicitly set in the `PublicFacingController`, as we:

1. Don't want admin pages iframed (default behaviour works here, and is
   overridden in Puppet anyway)
2. Want this for public requests only

- [Pivotal](https://www.pivotaltracker.com/n/projects/780127/stories/92019798)
- [Zendesk](https://govuk.zendesk.com/tickets/1001074)