Remove assertion for grey image
Some features were testing that an img was returned for people. Some
tests were set up without these appointments having an image so now they
are a white square with no alt tag. This commit removes these assertion
and in some tests the next assertion checks that there is a link to the
appointment gives us enough confidence that the page functions correctly.
For any tests which don't check the name I have changed them to do this.