When unpublishing, republish the draft edition to draft stack
We want the draft stack to have the current draft documents. Unpublishing an
edition changes the state of that edition to "draft" and sends an "unpublishing"
content item to publishing-api. Because publishing-api mirrors changes to live
to draft, this means that draft stack has the unpublishing, but we want it to
have the draft edition. Therefore, we need to explicitly republish the draft to
the draft stack.