Ensure correct role model class is instantiated
This fixes a bug that causes ministerial roles to be created without a
content_id, which makes it impossible to edit them as they fail validation.

There is model-specific code on the MinisterialRole class to generate
content_ids  and push them to the Publishing API. Because of the way that
the STI type column was being inferred and set from within the instance (with
the role_type= setter) rather than at the point it was instantiated[1],
ministerial roles were being created as base Role instances, which results in
the publishing API code being skipped.

[1] If you call Role.new(type: 'MinisterialRole'), you get an instance of
MinisterialRole back. If you set type after the fact on a Role instance, it
remains (and behaves like) an instance of Role until it is reloaded.