Remove policies from /get-involved
These were hard-coded, likely to break or become inaccurate
as changes are made to policy titles/slugs.

For users interested in policy, it's still surfaced in
the header nav.

Remove mention of policies on the how-gov-works page and
rejig the layout to avoid leaving an empty column. Follow
the pattern the `Legislation` follows.