Use SitewideSetting rather than using ENV vars
We want to be able to change policy flags without having the need to
update the puppet or deployment repos. The current approach of using ENV
variables would require this.

Use the sitewide setting module which can take a name and add boolean
value and a rake task to update or create those settings.

New sitewide settings can be created with:

    bundle exec rake sitewide_settings:set[new-key,false]

And if you want to update one you can call:

    bundle exec rake sitewide_settings:set[new-key,true]