Add temporary redirects for policies
This is necessary because until policies are moved to their new URLs,
both new and old slugs exist under the same content_id in the content
store, So any case study that links to them gets the new slug, which
points to a path that doesn't exist yet.

This will be removed when the new policies are ready to go live.