Only log header when base class changes.
Previously the Populator::ProgressLogger would log a new header each
time the class changed. But since we use STI, Editions can be one of
several classes in no particular order, meaning that a new header would
be logged almost every time. This change uses the base class if it
exists to determine whether to log a new header.