Remove policy-related subscription URLs in new-policy mode.
The `relevant_to_local_government` URLs will already
be trimmed at the moment due to the unique way
`#published_related_policies` is funded.

This is shortly going to be changed, hence this
guard condition.

Local government relevance will in future have to
be dealt with in the data pushed to the content
store in order for the newer emailing system to
handle it.