Rake task and CSV for policy redirects.
The rake task election:redirect_policies will register redirects for all
published policies currently on GOV.UK. It will do this using the generated CSV
file that contains the redirects. The script used to generate these redirects is
also included for reference, and in case the redirects need to be regenerated
due to any further changes to slugs of documents or policies.