Default publishing locales to :en.
If an object has not been created with fields that
are translated, globalize does not create the
translation associations.

With no translations it considers the object to
have no locales, which prevents our publishing API
callbacks from working properly.

This defaults the locales for this case to :en
without adding :en to e.g. [:fr]

It does not cover the case mentioned in the comments
regarding objects which are not translated at all.