Generate statistics feed URLs for email subscriptions
In order to send email notifications to people who have subscribed from
the statistics page we need to make sure extra feed URLs are generated
for any document types included on that page.

This change makes the `SubscriptionUrlGenerator` return URLs for the
statistics feed for statistics, national statistics, and statistical
data set documents. This in in addition to the publication feed URLs, as
these documents appear in both feeds.