Auto-update parts of how-gov-works for new gov
Some parts of the page are specific to the 2010-2015 government
and will need to be updated when a new government is formed. Even
if it is similar to the previous coalition many details may still
be inaccurate, so it's best to show a simplified version of the
page and then update it once there is more information about the
new governent.

Some sections, like policies, may not reflect the new govenrment
and/or may be updated, so hide those while they settle.

Equally, minister/department counts are likely to change during
the formation of a new government, as with a reshuffle, so hide
those to avoid confusion during the initial stages of a new gov.

This also covers specific descriptions of the government, like
who the priminister is, if there is a coalition. Takes the page
page to a very abstract description of the government.

This introduces quite a bit of tech debt to this page, but it will
be short lived and cleaned up by the election team as soon as
there is updated content to go on the page.