Government.current handles overlapping government
This fixes a bug where Government.on_date would return the previous government
if given the date where an old government closes and the new government is
formed. This would result in documents published on the day to be incorrectly
marked as historic.

This changes the factories for current and previous government to better reflect
the way they will be created in reality, i.e. with the current government
starting on the same day as the previous government closed.