Don't overwrite government end date when closing.
This allows GDS admins to set an end date ahead of
time but still use the "close the government"
functionality to close the associated ministerial
roles.

This would happen in a situation where the government
was being closed in whitehall admin on the day
after the actual closing (or later).  Unlikely, but
possible.