Ensure role appointment data integrity
Fixes a bug which causes the subsequent appointments to ministerial roles closed
when the government is closed to be invalid.

The RoleAppointment model does various checks to ensure appointments do not
overlap. This code assumes that all the role appointments have dates and not
times set for both the started_at and ended_at fields. This is fine when they
are being administered through the admin interface as the form fields only allow
dates to be selected. The bug is caused by the code that closes the government
and ends all existing ministerial roles with a time and not a date.

The precise problem is a little nuanced and is rooted in the fact that the
role_appointments started_at and ended_at columns are datetime and not date
columns. If an existing role appointment is closed with a timestamp
(e.g. 12th May 00:09:00), any role appointment that is created the same day
using the admin interface will have a started_at that is less than the ended_at
of the previous appointment because ActiveRecord is casting the date supplied to
a timestamp set to the beginning of the day (e.g. 12th May 00:00:00).
 # Please enter the commit message for your changes. Lines starting