Ensure feature-flag is cleaned up after features
DatabaseCleaner is set to use the :transaction strategy in cucumber tests. But
this only becomes active in a "Before" filter, which means that anything created
in an "Around" filter will not be included in the transaction (the order of the
filters is around begins -> before -> after -> around ends ). Therefor, anything
created in an Around filter should be cleaned up afterwards, otherwise it will
linger beyond the feature it was created in.