Index all people in search, including ministers
Previously: Ministerial roles are indexed in search, as well as
people, so to avoid someone with a ministerial role appearing
twice in search we didn't index people with ministerial roles.

We're not happy to include the duplicates, as they are slightly
different, and the 'people' result may be more useful for users
than the role.

The finding things team are happy with this change, and once
all people are being index they're going to look at weighting
to promote people results above roles.

We want all people to be in search so that other results, which
reference a perosn through a 'people' field (eg, an edition linked
to a minister), can have the 'people' field automatically expanded
to include more fields than just the slug, as we currently do for
some fields, like organisations