Update Rack to 1.6.1
This fixes a bug where all form fields in a request are counted against the multipart file limit:
https://github.com/rack/rack/pull/814

This is an issue in production - some publications and organisations have a large number of dynamically added nested forms. When the form is submitted, a 500 is returned because the number of fields exceeds the limit. It only affects forms posted with multipart, e.g. ones that contain file fields.

Zendesk: https://govuk.zendesk.com/agent/tickets/1037587
Story: https://www.pivotaltracker.com/story/show/94668330