Send policy groups to content store as placeholder working groups.
"Working group" is the new cross-gov.uk name for a
Whitehall PolicyGroup aka PolicyAdvisoryGroup.

The Whitehall Group model is separate (though
similar) and will eventually be merged, though both
are still in relatively active use.

These working groups need to be in the content
store so that policies (now published from the
Policy Publisher) can continue to be tagged to them.

This code adds content IDs to the policy groups
via a normal migration, and sends them to the
content store via a data migration as custom placeholders.

https://trello.com/c/pV7ExOoE/203-5-associate-policy-to-a-group