Remove descriptions from lists of policies
New style policies don't have short descriptions like the old ones used
to have. This means the pages don't look great with the descriptions
visible. Also we have now switched to showing more policies in these
lists so a shorter list is preferable.

Remove a `overflow:hidden` as it was previously used to contain floats.
The only element with the element which has any floated elements also
has our float clear fix so it was redundant and caused the top of some
of the text to be chopped off.