Don't create world location in global filter
A world location was being created for every single feature, when it is only
really needed for those that involve editing contacts. This makes the creation
of the dependent location explicit in those features that require it.

This was causing an exception to be raised because of cucumber is
evaluating features/support/env.rb before features/support/publishing_api.rb,
which means the calls to Publishing API have not been stubbed in time for the
create(:world_location) call. This was previously working because ActiveRecord
suppressed errors raised in after_commit callbacks prior to version 4.2.