Add script to rewrite migrated policy publication history
This script/rewrite_policy_publication_history.rb script will rewrite the
document history on publications that were generated to archive content from
the old Policy format.

It does the following:

	* Makes a note of the existing change history on the publication
	* Clears the publication change history so it starts from a blank-slate
	* For every major change on the original policy, re-edition the publication and
	use the change note and public timestamp of the policy major change
	* For those changes on the publication that were noted in the first step,
	re-edition the publication and adopting the public timestamp of those changes

This should result in a merge document history on the publication that
adopts the history of the original policy, followed by subsequent changes for
the transition to a publication and any major changes that may have been
performed since the publication was initial published.

The expects all publications to be in a "published" state, with no open "draft"
or "submitted" editions. Any remaining "draft" or "submitted" editions should be
either disgarded or published before running the script as otherwise it will
fail.

Note that although the script performs the history rewriting in a transaction,
it is not idempotent and so should only ever be run once, otherwise the history
will be duplicated. If the script exists with errors however, the transaction
will rollback and the script can be retried.