DocumentHistory handles adopted supporting pages
This fixes an issue that results in a strange document history on policies that
adopt previously-published supporting pages from other policies.

Policies can linked to supporting pages that have already been published with
another policy. In such a case, it is possible that the supporting page will
have been published before the policy itself. This can results in a strange
change history where the entry for the supporting page appears before that of
the policy's "First published." change. This updates the code such that any
change notes on supporting pages published before the policy itself are ignored,
resulting in a more accurate change history.