Fix sideways scrolling service priority org pages
The `.service-priority` class was including grid-row styles to contain
floats, but this came with a large negative margin. This negative
margin caused sideways scrolling on thin page widths.

Instead use the existing `floated-children` class on the inner block to
contain the floats (and alter padding at different screen sizes) and
stop extending with `grid-row`.

Outcome is a page that looks exactly the same but doesn’t sideways
scroll.