Add ordering to ministerial departments
There is a need to manually curate the order which departments are
listed on the ministers page. This adds a ministerial ordering to
organisations which can be updated on the cabinet ministers page.

Also updated the prefix on the ministers by org objects on the ministers
page as some ministers were appearing twice and causing the same html ID
to be used twice. As this is invalid markup nokogiri was failing.