Update auto-topic picker to use new style policies
Since the move to using content store policies the auto picking of
topics stopped working. This brings back the old functionality so that
when you pick a policy with a topic associated to it that topic will be
auto selected in the admin.

Removed the TopicPresenter and instead just used the topic id's as the
presenter as nothing else is needed by the JavaScript and the extra
layer added confusion as to what was rendered. Added a factory for
classification_policy so that the controller could be unit tested.