Add a continue button to role form
You currently create role appointments from the role edit page. Having a
continue button on the form means if you are creating a new role you
will be able to straight away create an appointment for that role rather
than having to re-find that role in the list of roles.