Send detailed format to Rummager
In order to get proper objects back from Rummager for `display_type` we
need to create a new field which will use the Rummager allowed values
feature to return the correct label. This will allow us to filter for
`detailed_format` within Rummager powered pages such as Finders.

All editions will need to be republished when this is live.

Ticket:
https://trello.com/c/TAuAS4wp/174-8-add-more-filter-options-to-policy-pages