Fix case-[in]sensitivity issue for fractions
As most devs [citation needed] use Macs with case-insensitive file
systems to work on Whitehall, we should probably not have two versions
of the same file with differently cased names if we can avoid it.

Notably, this breaks the Github for Mac client which is one of the
better and more popular OS X git guis. When you boot the app up it is
confused about seeing `1_X.png` and `1_Y.png`, but only `1_x.png` and
`1_y.png` in the filesystem, so marks the uppercase versions for
deletion.