Remove all Import-related code
The import feature was used during transition to mass-import documents from
transitioning websites. Since transition is over, it is no longer required and
can be removed.

To remove the importer, the behaviour of the speed tagger has had to be changed
slightly. Previously, it was possible to speed-tag an imported edition and
automatically be redirected to the next imported edition to be tagged. This was
done by finding the next document from the Import the first document was
imported in. Because the Import code does not exist anymore, it isn't possible
to find the next document. Instead, the user is redirected to the edition show
page.