Cache policies from content register
There are currently lots of calls to get the policies from
content-register to power some of the frontend pages. This is because we
don't store the policy names in whitehall so have to get the full list
to get the names to display metadata on pages.

Content-register is designed to be used only by publishing tools so
doesn't want a cache put in front of it. Eventually when the whitehall
frontend pages are moved into the content-store properly there won't be
any need to call content-register as the link expansion will be done in
content-store.

This adds a 5 minute cache onto the responses from content-register so
that whitehall isn't hitting it to frequently (up to many 10's of times
a second at the moment).