Stop unpublish when other pre-pub editions exist
If a document has a pre-publication edition you shouldn't be able to
unpublish the document. If you are able to you can get into the state
where there are two pre-publication documents and the system can get
very confused.

The exact replication which we have seen in the wild is:

- Take a published document
- Create a new draft of that document and submit it to 2nd eyes
- Unpublish the published document
- You now have two editions which are both pre-publication

If you publish these in the wrong order the code which depends upon the
edition with the greatest `id` will break. This change stops you from
ever getting into this state.

Update the check which looked for draft editions and make it more
generic to handle any pre-publication editions.