Don't use a dummy contentapi in development
I was trying to reproduce a bug with the integration in development but it
wasn't being triggered. I think this was added when Whitehall was kept separate
to the rest of the stack, to enable it to be worked on without needing the rest
of GOV.UK set up in development. Times have changed and we should keep things
simple and consistent with our other apps which use real integration in dev.

I have also removed the ability to specify a custom URL for the contentapi for
the same reasons - we expect people to use the real integration.