Remove spring.
It's enabled by default in dev, and leaves whitehall processes running
on the dev VM. The fact that it's enabled by default is somewhat
surprising. Additionally various developers use their own app preloaders
(eg zeus).

People who know about spring and want to use it can use it without it
needing to be added to the Gemfile[1]. It therefore makes sense to
remove it.

[1] https://github.com/rails/spring#use-without-adding-to-bundle