Fix broken fix for email notifications env indicators
The fix here: c74383ac0f5f4de916b33c969a9fb6e4cc6c7905 meant that emails in
non-production environments stopped being sent.

This seems to be because square brackets need to be escaped. The parentheses
that we used previously were treated as a comment and so not displayed to users
whereas the square brackets seem to prevent the email from being sent.

Emails in production were unaffected because they do not include the square
brackets.

I've used the mail gem because it will quote the 'display_name' as required but
it also reduces the knowledge our code has to have about formatting the 'from'
field correctly.