Data migration for adding an EU logo to a detailed guide
As GOV.UK,
we need to add the EU logo to entrance pages of content which reference EU funding,
so we help departments remain compliant.

This data migration adds an EU logo to the last edition of a particular document.
There is no editing interface for this field so it has to be done as a migration.
The code was copied from [the data migration of a previous pull request](https://github.com/alphagov/whitehall/pull/2164).