Rake task for adding logos to documents
As GOV.UK,
we need to add the EU logo to entrance pages of content which reference EU funding,
so we help departments remain compliant.

This rake task allows adding a logo to the last edition of a particular document.
The logo will be rendered for any document, but the CSS scaling only works correctly
for detailed guides and document collections. The main usecase is to add the EU funding
logo. The code was taken from [the data migration of a previous pull request](https://github.com/alphagov/whitehall/pull/2164).