Hide empty link from accessibility software
The `attachment_thumbnail` will always produce an image with an empty
`alt` attribute. This causes the link which wraps the thumbnail to
always contain no content. Accessibility software doesn't like empty
links. As the link wrapping the header points to the same place as the
link wrapping the thumbnail we can just hide the thumbnail using
`aria-hidden`. This also stops the possibility of having a duplicate
link (which is also non-ideal).