Use ERDF logo from asset-manager
In #2280 we made the detailed guides use the logo URL from the
database. Using assets from `app/assets` directly will cause problems
because there is no reference in the code to this image anymore, so the
image could be deleted in the future.

This commit changes the database URL to a static, never-to-be-deleted,
version uploaded via the asset-manager.