Expose analytics_identifiers in API responses
We want to consistently tag pages in Google Analytics with:
* organisations
* world_locations
* world_organisations

The natural value to send would be the path or slug, however the limit on the
length of a dimension value in Google Analytics combined with the fact that
content is often tagged with multiple organisations/locations/etc means that
these would be truncated. Instead, we send a short code - the
analytics_identifier. Users of Google Analytics then translate these codes back
into meaningful values.

Whitehall mints these identifiers, so for content not published and rendered in
Whitehall to send this data to Google Analytics, the identifiers need to be
exposed to the other apps. This can then be exposed in contentapi which in turn
the frontend can pass to Google Analytics (currently via Slimmer).

There will be a separate change to include these values in content-store.