Redirect some worldwide priorities
All worldwide priorities were withdrawn in an earlier data migration [1],
which left them displayed but with a banner showing that they were withdrawn.
We now need to redirect 23 of them to other pages on GOV.UK.

For other document formats we would be able to use the unwithdraw_edition rake
task to republish the documents, and then users would be able to re-withdraw
them differently to set up the redirects. However, the worldwide priority
format is deprecated and the code to handle them has already been removed from
the admin interface, making this approach impossible.

This data migration follows the pattern of
https://github.com/alphagov/whitehall/commit/69a4bd193cd0abffdd49e11ce813ebf98dac7a91
by modifying the existing unpublishing record to set up the redirect. We also
need to change the state of the edition back to 'draft' so that it isn't found
by find_document_or_edition on the controller [2], allowing it to fall back to
redirect [3] instead of rendering the page.

A few previous data migrations have used CSV files too, eg
https://github.com/alphagov/whitehall/commit/1fc1e1deaf9c941bb1fcc5b0c37458c771d6ad6c.

[1]: https://github.com/alphagov/whitehall/blob/286bf40509a1ecab55af3d1b606d5d0221bcd3fa/db/data_migration/20150427095111_archive_all_worldwide_priorities.rb
[2]: https://github.com/alphagov/whitehall/blob/286bf40509a1ecab55af3d1b606d5d0221bcd3fa/app/controllers/documents_controller.rb#L29
[3]: https://github.com/alphagov/whitehall/blob/286bf40509a1ecab55af3d1b606d5d0221bcd3fa/app/controllers/documents_controller.rb#L39