Gracefully handle really wide CSV attachments
A CSV was uploaded which had ~300 rows and 16,000 columns. When this file was
"viewed online" (aka "preview"), Whitehall would consume huge amounts of memory
trying to render the HTML until it was killed by unicornherder.

The CSV was fixed to solve that particular instance, but it shouldn't be
possible to break Whitehall by uploading and viewing a document.

We already truncate CSVs that are too long, let's truncate them when they're
too wide.