Check with url-arbiter before adding route
Before re-registering the route, check with url-arbiter that whitehall
owns the url. If it’s owned by another app, skip it. If url-arbiter
doesn’t know about it, reserve it for whitehall.

This uses a dummy API adapter for url-arbiter because adding the
govuk-client-url_arbiter gem for this one-shot script didn’t seem like
a good idea.