Always show works with on orgs list
If any organisation works with other organisations we should be showing
their works with line.

We currently have a collection of organisations which work with other
organisations who have links on their org homepage to their space in the
org list. This is so you can see the orgs they work with. However, we
weren't currently showing who those orgs work with as we were only
showing it for some org types. This lets those links take you to
meaningful information.