Send case study to draft when updating in all states
The `edition.draft?` guard in the `DraftEditionUpdater` service object
was too narrow. We should allow updates to editions in the `submitted`
or `rejected` state as well.

This problem was compounded by the fact that there wasn't sufficient
error handling around the call to `updater.perform!` in the
`Admin::EditionController`. The consequence was that the `#perform!`
call would silently fail and the edition would be updated, but the draft
would not be sent to the content preview environment.

I've tightend up the testing around this by adding a new functional test
mixin for use in edition admin controllers
`should_send_drafts_to_content_preview_environment_for`, and using this
in the admin case study controller test.

Trello:
https://trello.com/c/yWNA28st/295-bug-case-studies-in-2i-don-t-appear-in-content-preview