Bugfix: make edition#update work properly
We must assign attributes to the edition before calling
`DraftEditionUpdater#can_perform?`, otherwise it's checking against the
edition before the user input has been assigned to the model object.

However, we don't want to call `#save` before we've checked that the
updater is safe to perform.

In an ideal world, we'd push all of these steps down into the
`DraftEditionUpdater` service object. But due to the current api of
service objects there's no clean way to pass through the current user to
the service object (which is needed by the `#edit_as` method).

Therefore, I've split up the `#edit_as` method into two pieces. This
method is now defunct as it was only actually called in one place. So as
a bonus, I've deleted a bunch of unit tests of that method.