Amend search description when necessary
Currently, Rummager is in charge of appending the prefix to open organisations.
We want Whitehall to take care of its data.

The work in this pull request moves the responsibility of updating the search
description of Whitehall organisations to Whitehall itself.

In the below pull request we added a temporary guard in Rummager to avoid displaying
a repeated prefix in search results when this PR will have been merged and deployed.
After deploy, the guard will become redundant and we shall remove it.

https://github.com/alphagov/rummager/pull/482