Improve Ministerial Role search title and description
Currently the title of a ministerial role in search is "Person (Role)".
This causes confusion for users as the `person` pages have almost the
same title.

The original reason for "Person (Role)" was that Person pages weren't
being indexed, so this role page was the main search result for a
person in cabinet. We've started indexing people as well, so this title
doesn't really make sense anymore.