Tag organisations with self in search results
Currently organisation pages are not tagged with themselves as
organisations.

This causes odd behaviour when using facetted search: searching for
HMRC will show the organisation page as the first result, but the org
page disappears when selecting HMRC in the organisations filter.

Also tags the page with the parent organisation, so that child agencies
like the Valuation Office Agency are also included in search results
for HMRC.