Generate a CSV report of detailed guides
A rake task that outputs a report to assist in ensuring that Detailed
Guides are accessible at the correct URLs or are redirecting where
necessary.
It gathers data from Whitehall about the latest and published editions,
whether the slug has a “deleted-“ prefix and if an Unpublishing record
exists for it. It calls out to the Router API and reports on what route
exists for the slug at both the root and /guidance URLs. It also checks
with Panopticon about whether the document is duplicated, hasn’t had
its slug changed yet, is missing or doesn’t have the correct paths set.