Ensure withdrawn detailed guides are sent to panopticon
The `panopticon:register_guidance` rake task wasn’t sending `withdrawn`
Detailed Guides to panopticon. This is incorrect as Whitehall needs to
handle requests for withdrawn documents, and running this task to
register new namespaced routes was missing the withdrawn documents.

Using the `Document.published` scope filters by documents with editions
in the `published` or `withdrawn` states.