Push root redirects for moved detailed guides
Rake task that loops through all Detailed Guides, published or not, and
ensures the correct root route is set for it and any translations

If the guide has ever been published (`published` or `withdrawn`
state), add a redirect to the new `/guidance` URL. This will also
happen if an `Unpublishing` exists for the slug and it’s not marked as
an immediate redirect, as Whitehall needs to respond with the correct
“Gone” notice.
If an `Unpublishing` exists and is marked as an immediate redirect, or
was unpublished due to consolidation, add a hard redirect to the
target, meaning that the /guidance/:slug URL is no longer needed.

All redirects are created by publishing a redirect item to the content
store.