Send mainstream browse taggings to rummager
Panopticon allows users to tag edition-content to mainstream browse
pages. Tagging something to a mainstream browse page includes it in the
list of links on a /browse page (https://www.gov.uk/browse) such as
https://www.gov.uk/browse/business/imports-exports.

We'd like to make the collections application serve these lists of
links from rummager (https://github.com/alphagov/collections/pull/151)
instead of from content API. This works because publisher and
specialist-publisher already send the browse pages to which the item is
tagged to rummager. This allows us to filter with rummager:

https://www.gov.uk/api/search.json?filter_mainstream_browse_pages[]=busi
ness/imports-exports

Whitehall is the last application that doesn't send
`mainstream_browse_pages` taggings to rummager. It doesn't even know
about which browse pages are tagged to an edition, because tagging to
`/browse` is special and should be restricted to GDS. This is a bit
unfortunate, because it requires us to fetch the taggings from the
content api before we can send the document to rummager.

Trello: https://trello.com/c/7qw4TRiC