Improve sending browse pages to rummager
Currently whitehall asks content-api for the browse pages tagged to an
artefact. This was added in PR #2322.

However, it does so by asking the content-api for each edition during
indexing, about 130.000 times currently. This commit uses the new
endpoint on content-api that was added specially for this use-case:

https://github.com/alphagov/govuk_content_api/pull/226

This endpoint returns all artefacts tagged to mainstream browse pages,
so that we can do one request to get them all.

We use the standard Rails caching mechanism here to cache the response.

We extend `GdsApi::ContentApi` in this app because no other clients may
use the `artefacts_tagged_to_mainstream_browse_pages` endpoint. This is
a pattern used by the publishing API:

https://github.com/alphagov/publishing-api/blob/560112fc1019d77f4dce4f93
061165a574d5d7eb/lib/clients/content_store_writer.rb