Remove SassC
Currently in development if you change any of the Sass files which
aren't either `frontend.scss` or `admin.scss` the change will not get
picked up and you won't be able to see your style changes. This is due
to some excessive caching which only busts when there is a change to
those two files.

Until that can be properly fixed remove SassC from this repo.