Send Editions' content_id to gds-api-adapters
We want to use content IDs instead of slugs to represent items in curated
lists, so that things like slug changes don't break the linkage. In order to do
this, we need to ensure that all content in publisher has a content-id, and
these content-ids are published to the publishing API.

This newly added test check that RegisterableEdition has a content_id.

Part of:
https://trello.com/c/rmBoBPm9/290-ensure-that-all-content-in-publisher-has-a-content-id