Make content_id available in RegisterableEdition
We want to use content IDs instead of slugs to represent items in curated
lists, so that things like slug changes don't break the linkage. In order to do
this, we need to ensure that all content in publisher has a content-id, and
these content-ids are published to the publishing API.

This newly added test check that RegisterableEdition has a content_id.

This work does what I (wrongly) thought the work in this commit was doing: https://github.com/alphagov/whitehall/commit/7b8aeec7882e7de360974973f3c75be0b1e76ec4