Add beta label to farming
DEFRA wants a beta label on their new "farming" topical events page.
There currently is no way to do this. In the interest of speed, we
chose for a temporary variable on the TopicalEvent model.

If it turns out that more topical events need a beta label, we can
replace this method by a proper column in the table.

Trello: https://trello.com/c/nXsxrLLc