Rename remaining occurrences of "topics"
~3 months ago we renamed "topics" to "policy areas" but these instances have
been missed.

The error messages are dynamically generated, we therefore need to send the
message to :policy_area instead of :topic. I do not like this approach because
we're adding a new symbol to Edition's list of errors. This exception could be
removed by renaming "topics" to "policy areas" not only in the views but also
in the "backend".

Any idea on other approaches to fix this issue?