Upgrade Slimmer, set layout in standard way
This is a no-op change to app behaviour. The same Slimmer layout is used as be
before, but set in a simple, standardised way with `slimmer_template`, which is
easier to grep for than the other methods.

`slimmer_template` is available directly from `Slimmer`, so I've removed the WH
implementation. The template setting logic will stay the same for attachments,
but is slightly cleaner, and doesn't use parens, to be consistent with use of
`slimmer_template` elsewhere.

This is a pretty minor change, but it bring WH inline with other FE apps, so its
worth it for the consistency.

[1] https://github.com/alphagov/slimmer/pull/136