Style stat-headline markdown
Allow writers to pull out important numbers. This can be a single
number or a group of them.

“Departments are about to create and publish their Single Departmental
Plans. These will be published to GOV.UK as HTML publications.

It is a ministerial priority that depts can pull out important figures
periodically throughout the HTML publication in large font, and can
visually represent data in graphs (this is just images as per the
budget).”

https://trello.com/c/z4YmHWDp/127-html-pubs-add-markdown-to-allow-users-
to-add-large-font-and-graphs-medium