At medium (tablet) screen sizes, float the container instead of the image
Due to differences in template HTML, on tablets previous the image was being floated to the right side of the main copy. This generally works, but in templates with a figcaption this was being left on the original side above the content.

Now that we’ve normalised the templates we can apply the floating CSS to the figure rather then directly to the image, which keeps the figcaption directly below the image in this layout.